smart_construction_backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=1.4
pymysql
python-dotenv
passlib[bcrypt]
python-jose
requests
pydantic
